修改方案报价实体--存成一张表
