--- table/src/main/java/com/mengyunzhi/article/repository/DetailRepository.java
@@ -1,9 +1,0 @@
-package com.mengyunzhi.article.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-/**
- * Created by Mr Chen on 2017/8/30.
- */
-public interface DetailRepository extends CrudRepository<Detail,Long> {
-}
