+++ table/src/main/java/com/mengyunzhi/article/repository/Plan.java
@@ -9,19 +9,31 @@ public class Plan {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
+
     @OneToOne
     private Article article; //文章实体
+
     private Integer adultNum;   // 成人数
     private Integer childNum;   // 儿童数
     private String currency; // 币种
+    private String type; // 类型：机票，签证，旅游，保险
+    private DecimalFormat adultUnitPrice;   // 成人单价
+    private DecimalFormat childUnitPrice;   // 儿童单价
+    private DecimalFormat totalPrice; //总价
+    private String remark; // 备注
     private DecimalFormat totalCost; // 总费用
     private Date lastPayTime; // 最晚付款时间
 
-    public Plan(Article article, Integer adultNum, Integer childNum, String currency, DecimalFormat totalCost, Date lastPayTime) {
+    public Plan(Article article, Integer adultNum, Integer childNum, String currency, String type, DecimalFormat adultUnitPrice, DecimalFormat childUnitPrice, DecimalFormat totalPrice, String remark, DecimalFormat totalCost, Date lastPayTime) {
         this.article = article;
         this.adultNum = adultNum;
         this.childNum = childNum;
         this.currency = currency;
+        this.type = type;
+        this.adultUnitPrice = adultUnitPrice;
+        this.childUnitPrice = childUnitPrice;
+        this.totalPrice = totalPrice;
+        this.remark = remark;
         this.totalCost = totalCost;
         this.lastPayTime = lastPayTime;
     }
@@ -69,6 +81,46 @@ public class Plan {
         this.currency = currency;
     }
 
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public DecimalFormat getAdultUnitPrice() {
+        return adultUnitPrice;
+    }
+
+    public void setAdultUnitPrice(DecimalFormat adultUnitPrice) {
+        this.adultUnitPrice = adultUnitPrice;
+    }
+
+    public DecimalFormat getChildUnitPrice() {
+        return childUnitPrice;
+    }
+
+    public void setChildUnitPrice(DecimalFormat childUnitPrice) {
+        this.childUnitPrice = childUnitPrice;
+    }
+
+    public DecimalFormat getTotalPrice() {
+        return totalPrice;
+    }
+
+    public void setTotalPrice(DecimalFormat totalPrice) {
+        this.totalPrice = totalPrice;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
     public DecimalFormat getTotalCost() {
         return totalCost;
     }
@@ -93,6 +145,11 @@ public class Plan {
                 ", adultNum=" + adultNum +
                 ", childNum=" + childNum +
                 ", currency='" + currency + '\'' +
+                ", type='" + type + '\'' +
+                ", adultUnitPrice=" + adultUnitPrice +
+                ", childUnitPrice=" + childUnitPrice +
+                ", totalPrice=" + totalPrice +
+                ", remark='" + remark + '\'' +
                 ", totalCost=" + totalCost +
                 ", lastPayTime=" + lastPayTime +
                 '}';
